# banksampah_restful_api-openapi.yaml
openapi: "3.0.3"
info:
  title: "banksampah_restful_api"
  description: "API Spec Untuk Aplikasi Bank Sampah (BS) Mobile"
  version: "1.0.0"
servers:
  - url: "http://127.0.0.1:8080"
    description: "Local server"
paths:
...
  /api/accounts:
    post:
      tags:
        - "Account"
      summary: "Create Account"
      operationId: "createAccount"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                  example: "Nur Ikhsan"
                email:
                    type: "string"
                    example: "nurikhsan.0203@gmail.com"
                address:
                    type: "string"
                    example: "Jl. Taman Siswa No. 1"
                phoneNumber:
                    type: "string"
                    example: "081234567890"
                password:
                    type: "string"
                    example: "123456"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        format: "int32"
                        example: 1
                      name:
                        type: "string"
                        example: "Nur Ikhsan"
                      email:
                        type: "string"
                        example: "nurikhsan.0203@gmail.com"
                      address:
                        type: "string"
                        example: "Jl. Taman Siswa No. 1"
                      phoneNumber:
                        type: "string"
                        example: "081234567890"
                      password:
                        type: "string"
                        example: "123456"
  /api/login:
    post:
      tags:
        - "Account"
      summary: "Login to Application"
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                email:
                  type: "string"
                  example: "nurikhsan.0203@gmail.com"
                password:
                    type: "string"
                    example: "123456"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        format: "int32"
                        example: 1
                      name:
                        type: "string"
                        example: "Nur Ikhsan"
                      email:
                        type: "string"
                        example: "nurikhsan.0203@gmail.com"
                      address:
                        type: "string"
                        example: "Jl. Taman Siswa No. 1"
                      phoneNumber:
                        type: "string"
                        example: "081234567890"
                      password:
                        type: "string"
                        example: "123456"
  /api/signup:
    post:
      tags:
        - "Account"
      summary: "Sign Up on The Application"
      operationId: "signUp"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                  example: "Nur Ikhsan"
                email:
                  type: "string"
                  example: "nurikhsan.0203@gmail.com"
                password:
                  type: "string"
                  example: "123456"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "string"
                    example: "Berhasil"
  /api/accounts/{id_account}:
    get:
      tags:
        - "Account"
      summary: "Get One Account"
      operationId: "getAccount"
      parameters:
        - name: "id_account"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        format: "int32"
                        example: 1
                      name:
                        type: "string"
                        example: "Nur Ikhsan"
                      email:
                        type: "string"
                        example: "nurikhsan.0203@gmail.com"
                      address:
                        type: "string"
                        example: "Jl. Taman Siswa No. 1"
                      phoneNumber:
                        type: "string"
                        example: "081234567890"
                      password:
                        type: "string"
                        example: "123456"
    put:
      tags:
        - "Account"
      summary: "Update Account"
      operationId: "updateAccount"
      parameters:
        - name: "id_account"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      requestBody:
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                  example: "Nur Ikhsan"
                email:
                  type: "string"
                  example: "nurikhsan.0203@gmail.com"
                address:
                  type: "string"
                  example: "Jl. Taman Siswa No. 1"
                phoneNumber:
                  type: "string"
                  example: "081234567890"
                password:
                  type: "string"
                  example: "123456"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        format: "int32"
                        example: 1
                      name:
                        type: "string"
                        example: "Nur Ikhsan"
                      email:
                        type: "string"
                        example: "nurikhsan.0203@gmail.com"
                      address:
                        type: "string"
                        example: "Jl. Taman Siswa No. 1"
                      phoneNumber:
                        type: "string"
                        example: "081234567890"
                      password:
                        type: "string"
                        example: "123456"
    delete:
      tags:
        - "Account"
      summary: "Delete Account"
      operationId: "deleteAccount"
      parameters:
        - name: "id_account"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int32"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    format: "int32"
                    example: 200
                  status:
                    type: "string"
                    example: "OK"
                  data:
                    type: "integer"
                    format: "int32"
                    example: 1
...